fix: split setup page into two components
