+++ lib/games/sorting/getPhonemes.ts
@@ -4,6 +4,7 @@ const getPhonemes = () =>
   fetchServer<{ id: number; name: string }[]>(
     "/sorting_game/phoneme_options/"
   ).then((res) => res.json());
+
 export const getSpecificPhonemes = (phonemeIds: readonly number[]) =>
   getPhonemes().then((phonemes) =>
     phonemes.filter((phoneme) => phonemeIds.includes(phoneme.id))
